+++ rust/Cargo.toml
@@ -36,6 +36,7 @@ getrandom = { version = "0.2.3", features = ["js"] }
 rand = "0.8.5"
 schemars = "0.8.8"
 serde = { version = "1.0", features = ["derive"] }
+fraction = "0.10.0"
 base64 = "0.13"
 
 # non-wasm
