+++ rust/src/ledger/byron/witness.rs
@@ -1,0 +1,47 @@
+#[cfg(not(all(target_arch = "wasm32", not(target_os = "emscripten"))))]
+use noop_proc_macro::wasm_bindgen;
+
+#[cfg(all(target_arch = "wasm32", not(target_os = "emscripten")))]
+use wasm_bindgen::prelude::*;
+
+use crate::{crypto::{TransactionHash, LegacyDaedalusPrivateKey, BootstrapWitness, Bip32PublicKey, Vkey, Bip32PrivateKey, Ed25519Signature}, address::ByronAddress};
+
+#[wasm_bindgen]
+pub fn make_daedalus_bootstrap_witness(
+    tx_body_hash: &TransactionHash,
+    addr: &ByronAddress,
+    key: &LegacyDaedalusPrivateKey,
+) -> BootstrapWitness {
+    let chain_code = key.chaincode();
+
+    let pubkey = Bip32PublicKey::from_bytes(key.0.to_public().as_ref()).unwrap();
+    let vkey = Vkey::new(&pubkey.to_raw_key());
+    let signature = Ed25519Signature::from_bytes(key.0.sign(&tx_body_hash.to_bytes()).as_ref().to_vec()).unwrap();
+
+    BootstrapWitness::new(
+        &vkey,
+        &signature,
+        chain_code,
+        addr.attributes(),
+    )
+}
+
+#[wasm_bindgen]
+pub fn make_icarus_bootstrap_witness(
+    tx_body_hash: &TransactionHash,
+    addr: &ByronAddress,
+    key: &Bip32PrivateKey,
+) -> BootstrapWitness {
+    let chain_code = key.chaincode();
+
+    let raw_key = key.to_raw_key();
+    let vkey = Vkey::new(&raw_key.to_public());
+    let signature = raw_key.sign(&tx_body_hash.to_bytes());
+
+    BootstrapWitness::new(
+        &vkey,
+        &signature,
+        chain_code,
+        addr.attributes(),
+    )
+}
