+++ rust/src/ledger/common/binary.rs
@@ -1,0 +1,246 @@
+use std::io::{BufRead, Seek, Write};
+
+use cbor_event::{de::Deserializer, se::Serializer};
+use cbor_event::Special as CBORSpecial;
+use cbor_event::Type as CBORType;
+
+use crate::error::{DeserializeError, DeserializeFailure};
+
+// JsError can't be used by non-wasm targets so we use this macro to expose
+// either a DeserializeError or a JsError error depending on if we're on a
+// wasm or a non-wasm target where JsError is not available (it panics!).
+// Note: wasm-bindgen doesn't support macros inside impls, so we have to wrap these
+//       in their own impl and invoke the invoke the macro from global scope.
+// TODO: possibly write s generic version of this for other usages (e.g. PrivateKey, etc)
+#[macro_export]
+macro_rules! from_bytes {
+    // Custom from_bytes() code
+    ($name:ident, $data: ident, $body:block) => {
+        // wasm-exposed JsError return - JsError panics when used outside wasm
+        #[cfg(all(target_arch = "wasm32", not(target_os = "emscripten")))]
+        #[wasm_bindgen]
+        impl $name {
+            pub fn from_bytes($data: Vec<u8>) -> Result<$name, JsError> {
+                Ok($body?)
+            }
+        }
+        // non-wasm exposed DeserializeError return
+        #[cfg(not(all(target_arch = "wasm32", not(target_os = "emscripten"))))]
+        impl $name {
+            pub fn from_bytes($data: Vec<u8>) -> Result<$name, DeserializeError> $body
+        }
+    };
+    // Uses Deserialize trait to auto-generate one
+    ($name:ident) => {
+        from_bytes!($name, bytes, {
+            let mut raw = Deserializer::from(std::io::Cursor::new(bytes));
+            Self::deserialize(&mut raw)
+        });
+    };
+}
+
+// There's no need to do wasm vs non-wasm as this call can't fail but
+// this is here just to provide a default Serialize-based impl
+// Note: Once again you can't use macros in impls with wasm-bindgen
+//       so make sure you invoke this outside of one
+#[macro_export]
+macro_rules! to_bytes {
+    ($name:ident) => {
+        #[wasm_bindgen]
+        impl $name {
+            pub fn to_bytes(&self) -> Vec<u8> {
+                let mut buf = Serializer::new_vec();
+                self.serialize(&mut buf).unwrap();
+                buf.finalize()
+            }
+        }
+    }
+}
+
+#[macro_export]
+macro_rules! to_from_bytes {
+    ($name:ident) => {
+        to_bytes!($name);
+        from_bytes!($name);
+    }
+}
+
+#[macro_export]
+macro_rules! to_from_json {
+    ($name:ident) => {
+        #[wasm_bindgen]
+        impl $name {
+            pub fn to_json(&self) -> Result<String, JsError> {
+                serde_json::to_string_pretty(&self)
+                    .map_err(|e| JsError::from_str(&format!("to_json: {}", e)))
+            }
+
+            #[cfg(all(target_arch = "wasm32", not(target_os = "emscripten")))]
+            pub fn to_js_value(&self) -> Result<JsValue, JsError> {
+                JsValue::from_serde(&self)
+                    .map_err(|e| JsError::from_str(&format!("to_js_value: {}", e)))
+            }
+
+            pub fn from_json(json: &str) -> Result<$name, JsError> {
+                serde_json::from_str(json)
+                    .map_err(|e| JsError::from_str(&format!("from_json: {}", e)))
+            }
+        }
+    }
+}
+
+const BOUNDED_BYTES_CHUNK_SIZE: usize = 64;
+
+pub (crate) fn write_bounded_bytes<'se, W: Write>(serializer: &'se mut Serializer<W>, bytes: &[u8]) -> cbor_event::Result<&'se mut Serializer<W>> {
+    if bytes.len() <= BOUNDED_BYTES_CHUNK_SIZE {
+        serializer.write_bytes(bytes)
+    } else {
+        // to get around not having access from outside the library we just write the raw CBOR indefinite byte string code here
+        serializer.write_raw_bytes(&[0x5f])?;
+        for chunk in bytes.chunks(BOUNDED_BYTES_CHUNK_SIZE) {
+            serializer.write_bytes(chunk)?;
+        }
+        serializer.write_special(CBORSpecial::Break)
+    }
+}
+
+pub (crate) fn read_bounded_bytes<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<Vec<u8>, DeserializeError> {
+    use std::io::Read;
+    let t = raw.cbor_type()?;
+    if t != CBORType::Bytes {
+        return Err(cbor_event::Error::Expected(CBORType::Bytes, t).into());
+    }
+    let (len, len_sz) = raw.cbor_len()?;
+    match len {
+        cbor_event::Len::Len(_) => {
+            let bytes = raw.bytes()?;
+            if bytes.len() > BOUNDED_BYTES_CHUNK_SIZE {
+                return Err(DeserializeFailure::OutOfRange{
+                    min: 0,
+                    max: BOUNDED_BYTES_CHUNK_SIZE,
+                    found: bytes.len(),
+                }.into());
+            }
+            Ok(bytes)
+        },
+        cbor_event::Len::Indefinite => {
+            // this is CBOR indefinite encoding, but we must check that each chunk
+            // is at most 64 big so we can't just use cbor_event's implementation
+            // and check after the fact.
+            // This is a slightly adopted version of what I made internally in cbor_event
+            // but with the extra checks and not having access to non-pub methods.
+            let mut bytes = Vec::new();
+            raw.advance(1 + len_sz)?;
+            // TODO: also change this + check at end of loop to the following after we update cbor_event
+            //while raw.cbor_type()? != CBORType::Special || !raw.special_break()? {
+            while raw.cbor_type()? != CBORType::Special {
+                let chunk_t = raw.cbor_type()?;
+                if chunk_t != CBORType::Bytes {
+                    return Err(cbor_event::Error::Expected(CBORType::Bytes, chunk_t).into());
+                }
+                let (chunk_len, chunk_len_sz) = raw.cbor_len()?;
+                match chunk_len {
+                    // TODO: use this error instead once that PR is merged into cbor_event
+                    //cbor_event::Len::Indefinite => return Err(cbor_event::Error::InvalidIndefiniteString.into()),
+                    cbor_event::Len::Indefinite => return Err(cbor_event::Error::CustomError(String::from("Illegal CBOR: Indefinite string found inside indefinite string")).into()),
+                    cbor_event::Len::Len(len) => {
+                        if chunk_len_sz > BOUNDED_BYTES_CHUNK_SIZE {
+                            return Err(DeserializeFailure::OutOfRange{
+                                min: 0,
+                                max: BOUNDED_BYTES_CHUNK_SIZE,
+                                found: chunk_len_sz,
+                            }.into());
+                        }
+                        raw.advance(1 + chunk_len_sz)?;
+                        raw
+                            .as_mut_ref()
+                            .by_ref()
+                            .take(len)
+                            .read_to_end(&mut bytes)
+                            .map_err(|e| cbor_event::Error::IoError(e))?;
+                    }
+                }
+            }
+            if raw.special()? != CBORSpecial::Break {
+                return Err(DeserializeFailure::EndingBreakMissing.into());
+            }
+            Ok(bytes)
+        },
+    }
+
+}
+
+// we use the cbor_event::Serialize trait directly
+
+// This is only for use for plain cddl groups who need to be embedded within outer groups.
+pub (crate) trait SerializeEmbeddedGroup {
+    fn serialize_as_embedded_group<'a, W: Write + Sized>(
+        &self,
+        serializer: &'a mut Serializer<W>,
+    ) -> cbor_event::Result<&'a mut Serializer<W>>;
+}
+
+// same as cbor_event::de::Deserialize but with our DeserializeError
+pub trait Deserialize {
+    fn deserialize<R: BufRead + Seek>(
+        raw: &mut Deserializer<R>,
+    ) -> Result<Self, DeserializeError> where Self: Sized;
+}
+
+// auto-implement for all cbor_event Deserialize implementors
+impl<T: cbor_event::de::Deserialize> Deserialize for T {
+    fn deserialize<R: BufRead + Seek>(raw: &mut Deserializer<R>) -> Result<T, DeserializeError> {
+        T::deserialize(raw).map_err(|e| DeserializeError::from(e))
+    }
+}
+
+// This is only for use for plain cddl groups who need to be embedded within outer groups.
+pub trait DeserializeEmbeddedGroup {
+    fn deserialize_as_embedded_group<R: BufRead + Seek>(
+        raw: &mut Deserializer<R>,
+        len: cbor_event::Len,
+    ) -> Result<Self, DeserializeError> where Self: Sized;
+}
+
+pub struct CBORReadLen {
+    deser_len: cbor_event::Len,
+    read: u64,
+}
+
+impl CBORReadLen {
+    pub fn new(len: cbor_event::Len) -> Self {
+        Self {
+            deser_len: len,
+            read: 0,
+        }
+    }
+
+    // Marks {n} values as being read, and if we go past the available definite length
+    // given by the CBOR, we return an error.
+    pub fn read_elems(&mut self, count: usize) -> Result<(), DeserializeFailure> {
+        match self.deser_len {
+            cbor_event::Len::Len(n) => {
+                self.read += count as u64;
+                if self.read > n {
+                    Err(DeserializeFailure::DefiniteLenMismatch(n, None))
+                } else {
+                    Ok(())
+                }
+            },
+            cbor_event::Len::Indefinite => Ok(()),
+        }
+    }
+
+    pub fn finish(&self) -> Result<(), DeserializeFailure> {
+        match self.deser_len {
+            cbor_event::Len::Len(n) => {
+                if self.read == n {
+                    Ok(())
+                } else {
+                    Err(DeserializeFailure::DefiniteLenMismatch(n, Some(self.read)))
+                }
+            },
+            cbor_event::Len::Indefinite => Ok(()),
+        }
+    }
+}
