+++ rust/src/metadata.rs
@@ -1,3 +1,5 @@
+use crate::ledger::common::binary::{CBORReadLen, Deserialize};
+
 use super::*;
 use linked_hash_map::LinkedHashMap;
 
@@ -488,9 +490,9 @@ pub fn encode_json_value_to_metadatum(value: serde_json::Value, schema: Metadata
     use serde_json::Value;
     fn encode_number(x: serde_json::Number) -> Result<TransactionMetadatum, JsError> {
         if let Some(x) = x.as_u64() {
-            Ok(TransactionMetadatum::new_int(&Int::new(&utils::to_bignum(x))))
+            Ok(TransactionMetadatum::new_int(&Int::new(&to_bignum(x))))
         } else if let Some(x) = x.as_i64() {
-            Ok(TransactionMetadatum::new_int(&Int::new_negative(&utils::to_bignum(-x as u64))))
+            Ok(TransactionMetadatum::new_int(&Int::new_negative(&to_bignum(-x as u64))))
         } else {
             Err(JsError::from_str("floats not allowed in metadata"))
         }
